Add AMO autodeployment
Added script that wraps web-ext's sign command to succeed on submission.

Details:

- It succeeds (exitcode=0) when `has been submitted for review` appears in the output
- It leaves the output untouched and fails (exitcode=1) if it doesn't appear
- It drops the confusing error from the output on success

Code comments:

1. `s/\($OK\).*$/\1/` will drop `It passed validation but...`
2. `1,/$OK/p` will drop the `FAIL` and its following error
3. `perl` will error when `has been submitted` does not appear.

It previously used `ack --passthru --nocolor "$OK"` for the third step but it required `ack` 2.x (not on travis and not on `apt-get`). I had also tried to replicate the behavior with `sed` and `grep` but no luck. The hard part was outputting everything, always, and throwing an error at the same time, without generating more files or lengthening the script.