Removes box from no dependencies text
Installed this extension and it looks great. The only issue was this
seemingly broken unicode symbol rendering as a box with the "No
Dependencies" text. Not sure if this renders as emoji or something else
for other users, but on Arch Linux in Google Chrome it appears as a
rectangle and looks broken.